--- src/features/Account/AccountCreation.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import { RouteProps } from 'react-router';
-import ManageAccount, { ManageAccountModes } from './AccountManagement';
-
-const CreateAccount = (props: RouteProps) => {
-  return <ManageAccount mode={ManageAccountModes.CREATE} {...props} />;
-};
-
-export default CreateAccount;
