--- src/features/Account/AccountEdition.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import ManageAccount, { ManageAccountModes } from './AccountManagement';
-
-const EditAccount = () => <ManageAccount mode={ManageAccountModes.EDIT} />;
-
-export default EditAccount;
