+++ src/features/Account/ConfirmationEmailSentComponent.tsx
