--- src/features/Application/ApplicationCreation.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import ManageApplication, {
-  ManageApplicationModes,
-} from './ApplicationManagement';
-
-const CreateApplication = () => (
-  <ManageApplication mode={ManageApplicationModes.CREATE} />
-);
-
-export default CreateApplication;
