+++ src/features/Application/ManageApplicationContainer.tsx
@@ -37,8 +37,8 @@ import { Account, APIResponse, Hacker } from '../../api';
 
 import ValidationErrorGenerator from '../../shared/Form/validationErrorGenerator';
 
-import ResumeComponent from './Resume';
-import SchoolComponent from './SchoolSelect';
+import ResumeComponent from './ResumeComponent';
+import SchoolComponent from './SchoolComponent';
 
 import WithToasterContainer from '../../shared/HOC/withToaster';
