+++ src/features/Application/SchoolComponent.tsx
